+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `print_url` option to DataDragon class. defaults to False
 - Added 429 error rate limit exceeded handling in `_base_data_dragon_client.py`
 - Added error handling to `_get_latest_version` in `_base_data_dragon_client.py`
+- Added replays_by_puuid endpoint in `match.py`
 
 ### Changed